Happy Yestarë!
Added more specific notes about how the Calendar of Imladris
related to the Shire Calendar at the end of the Third Age.
Added Mythlore and Beyond Bree links.
Added a reference to my Beyond Bree article.
Added a reference for the Lunar Readalong T.A. 3019 Astron 27-28 events.
Updates and corrections to Deficit notes and rounding errors in results.